requirements.txt
biopython   # -- Bio.PDB for PDB api
requests    # -- requests for HTTP api (rfam rest api)